move `WindowMainStateListener` to separate file
